+++ R/file.R
@@ -6,19 +6,23 @@
 #'
 #' @param infile file to be encrypted
 #' @param outfile Non-existant file where the encrypted output is to be written
+#' @param key For symmetric encryption, 'pkey' should be NULL (default) and
+#'   'key' can be either a string (Default is 'pass') or a raw object. For
+#'   asymmetric encryption, both 'key' (private key of the encrypter) and 'pkey'
+#'   (public key of the decrypter) should be raw objects.
+#' @param pkey See 'key'
 #' @param ascii \code{TRUE} if the outfile is to be encrypted as a ascii file.
 #'   Default is \code{FALSE}
-#' @param key A string without embbeded NULL. Default is 'pass'.
-#' @param method Currently, a stub. It should be 'symmetric'(default)
 #'
 #' @return An invisible TRUE
 #'
 #' @examples
+#' # symmetric case:
 #' write.table(iris, "iris.csv")
 #' all(
-#'   encrypt_file("iris.csv", "iris_encrypted.bin")
+#'   encrypt_file("iris.csv", outfile = "iris_encrypted.bin")
 #'   , file.exists("iris_encrypted.bin")
-#'   , decrypt_file("iris_encrypted.bin", "iris_2.csv")
+#'   , decrypt_file("iris_encrypted.bin", outfile = "iris_2.csv")
 #'   , file.exists("iris_2.csv")
 #'   , tools::md5sum("iris_2.csv") == tools::md5sum("iris.csv")
 #'   , unlink("iris.csv") == 0
@@ -28,9 +32,9 @@
 #'
 #' write.table(iris, "iris.csv")
 #' all(
-#'   encrypt_file("iris.csv", "iris_encrypted.txt", ascii = TRUE)
+#'   encrypt_file("iris.csv", outfile = "iris_encrypted.txt", ascii = TRUE)
 #'   , file.exists("iris_encrypted.txt")
-#'   , decrypt_file("iris_encrypted.txt", "iris_2.csv", ascii = TRUE)
+#'   , decrypt_file("iris_encrypted.txt", outfile = "iris_2.csv", ascii = TRUE)
 #'   , file.exists("iris_2.csv")
 #'   , tools::md5sum("iris_2.csv") == tools::md5sum("iris.csv")
 #'   , unlink("iris.csv") == 0
@@ -38,24 +42,46 @@
 #'   , unlink("iris_encrypted.txt") == 0
 #' )
 #'
+#' # asymmetric case:
+#' alice <- keypair()
+#' bob   <- keypair()
+#' write.table(iris, "iris.csv")
+#' all(
+#'   encrypt_file("iris.csv", alice$private_key, bob$public_key, outfile = "iris_encrypted.bin")
+#'   , file.exists("iris_encrypted.bin")
+#'   , decrypt_file("iris_encrypted.bin", bob$private_key, alice$public_key, outfile = "iris_2.csv")
+#'   , file.exists("iris_2.csv")
+#'   , tools::md5sum("iris_2.csv") == tools::md5sum("iris.csv")
+#'   , unlink("iris.csv") == 0
+#'   , unlink("iris_2.csv") == 0
+#'   , unlink("iris_encrypted.bin") == 0
+#' )
+#'
 #' @export
 #'
 encrypt_file   <- function(infile
-                           , outfile
-                           , ascii  = FALSE
-                           , key    = "pass"
-                           , method = "symmetric"){
+                           , key     = "pass"
+                           , pkey    = NULL
+                           , ascii   = FALSE
+                           , outfile){
+
   # assertions            ----
   assert_that(is.string(infile))
   assert_that(file.exists(infile))
   infile <- try(file(infile, "rb"), silent = TRUE)
   if(is.error(infile)){
-    stop("Unable to read (possibly permission problem) binary file: ", infile)
+    stop("Unable to read (possibly permission problem) file: ", infile)
   } else {
     on.exit(close(infile), add = TRUE)
   }
-  assert_that(is.string(key))
-  assert_that(method %in% c("symmetric"))
+  if(is.null(pkey)){
+    method <- "symmetric"
+    assert_that(is.string(key) || is.raw(key))
+  } else {
+    method <- "asymmetric"
+    assert_that(is.raw(key))
+    assert_that(is.raw(pkey))
+  }
   assert_that(is.string(outfile))
   assert_that(!file.exists(outfile))
   if(ascii){
@@ -77,7 +103,12 @@ encrypt_file   <- function(infile
     stop("Unable to read from the connection or file. Ensure that connection or file is a open readable binary connection.")
   }
 
-  keyAsRaw <- try(hash(charToRaw(key)), silent = TRUE)
+  if(is.string(key)){
+    keyAsRaw <- try(hash(charToRaw(key)), silent = TRUE)
+  } else {
+    keyAsRaw <- key
+  }
+
   if(is.error(keyAsRaw)){
     stop("Unable to convert 'key' into raw. Possibly encountered an embedded NULL.")
   }
@@ -86,6 +117,12 @@ encrypt_file   <- function(infile
   if(method == "symmetric"){
     string_enc_raw <- data_encrypt(charToRaw(string)
                                    , keyAsRaw
+                                   , hash(charToRaw("nounce"), size = 24)
+                                   )
+  } else {
+    string_enc_raw <- auth_encrypt(charToRaw(string)
+                                   , key
+                                   , pkey
                                    , hash(charToRaw("nounce"), size = 24)
                                    )
   }
@@ -106,28 +143,31 @@ encrypt_file   <- function(infile
   return(invisible(TRUE))
 }
 
-#' @title Decrypt a connection or file
+#' @title Decrypt a file
 #'
-#' @aliases decrypt_file
-#' @description \code{decrypt_file}) decrypts a binary or a text file back to a
-#'   file.
+#' @description \code{encrypt_file}) encrypts a file as a binary or a ascii
+#'   file. \code{decrypt_file}) decrypts a text or a binary file (encrypted by
+#'   \code{encrypt_file})
 #'
-#' @param infile Encrypted file.
-#' @param outfile A non-existant file where the decrypted output will be
-#'   written.
-#' @param ascii \code{TRUE} if the outfile is a ascii file. Default is
-#'   \code{FALSE}
-#' @param key A string without embbeded NULL. Default is 'pass'.
-#' @param method Currently, a stub. It should be 'symmetric'(default).
+#' @param infile file to be decrypted
+#' @param outfile Non-existant file where the decrypted output is to be written
+#' @param key For symmetric decryption, 'pkey' should be NULL (default) and
+#'   'key' can be either a string (Default is 'pass') or a raw object. For
+#'   asymmetric decryption, both 'key' (private key of the decrypter) and 'pkey'
+#'   (public key of the encrypter) should be raw objects.
+#' @param pkey See 'key'
+#' @param ascii \code{TRUE} if the outfile is to be decrypted as a ascii file.
+#'   Default is \code{FALSE}
 #'
 #' @return An invisible TRUE
 #'
 #' @examples
+#' # symmetric case:
 #' write.table(iris, "iris.csv")
 #' all(
-#'   encrypt_file("iris.csv", "iris_encrypted.bin")
+#'   encrypt_file("iris.csv", outfile = "iris_encrypted.bin")
 #'   , file.exists("iris_encrypted.bin")
-#'   , decrypt_file("iris_encrypted.bin", "iris_2.csv")
+#'   , decrypt_file("iris_encrypted.bin", outfile = "iris_2.csv")
 #'   , file.exists("iris_2.csv")
 #'   , tools::md5sum("iris_2.csv") == tools::md5sum("iris.csv")
 #'   , unlink("iris.csv") == 0
@@ -137,9 +177,9 @@ encrypt_file   <- function(infile
 #'
 #' write.table(iris, "iris.csv")
 #' all(
-#'   encrypt_file("iris.csv", "iris_encrypted.txt", ascii = TRUE)
+#'   encrypt_file("iris.csv", outfile = "iris_encrypted.txt", ascii = TRUE)
 #'   , file.exists("iris_encrypted.txt")
-#'   , decrypt_file("iris_encrypted.txt", "iris_2.csv", ascii = TRUE)
+#'   , decrypt_file("iris_encrypted.txt", outfile = "iris_2.csv", ascii = TRUE)
 #'   , file.exists("iris_2.csv")
 #'   , tools::md5sum("iris_2.csv") == tools::md5sum("iris.csv")
 #'   , unlink("iris.csv") == 0
@@ -147,13 +187,28 @@ encrypt_file   <- function(infile
 #'   , unlink("iris_encrypted.txt") == 0
 #' )
 #'
+#' # asymmetric case:
+#' alice <- keypair()
+#' bob   <- keypair()
+#' write.table(iris, "iris.csv")
+#' all(
+#'   encrypt_file("iris.csv", alice$private_key, bob$public_key, outfile = "iris_encrypted.bin")
+#'   , file.exists("iris_encrypted.bin")
+#'   , decrypt_file("iris_encrypted.bin", bob$private_key, alice$public_key, outfile = "iris_2.csv")
+#'   , file.exists("iris_2.csv")
+#'   , tools::md5sum("iris_2.csv") == tools::md5sum("iris.csv")
+#'   , unlink("iris.csv") == 0
+#'   , unlink("iris_2.csv") == 0
+#'   , unlink("iris_encrypted.bin") == 0
+#' )
+#'
 #' @export
 #'
 decrypt_file   <- function(infile
-                           , outfile
-                           , ascii  = FALSE
-                           , key    = "pass"
-                           , method = "symmetric"){
+                           , key     = "pass"
+                           , pkey    = NULL
+                           , ascii   = FALSE
+                           , outfile){
 
   # assertions ----
   assert_that(is.string(infile))
@@ -171,9 +226,15 @@ decrypt_file   <- function(infile
   } else {
     on.exit(close(infile), add = TRUE)
   }
-  assert_that(is.string(key))
-  assert_that(method %in% c("symmetric"))
 
+  if(is.null(pkey)){
+    method <- "symmetric"
+    assert_that(is.string(key) || is.raw(key))
+  } else {
+    method <- "asymmetric"
+    assert_that(is.raw(key))
+    assert_that(is.raw(pkey))
+  }
   assert_that(is.string(outfile))
   assert_that(!file.exists(outfile))
   outfile <- try(file(outfile, "wb"), silent = TRUE)
@@ -207,22 +268,32 @@ decrypt_file   <- function(infile
     }
   }
 
-  keyAsRaw <- try(hash(charToRaw(key)), silent = TRUE)
+  if(is.string(key)){
+    keyAsRaw <- try(hash(charToRaw(key)), silent = TRUE)
+  } else {
+    keyAsRaw <- key
+  }
   if(is.error(keyAsRaw)){
     stop("Unable to convert 'key' into raw. Possibly encountered an embedded NULL.")
   }
 
   # decrypt    ----
   if(method == "symmetric"){
-    string_dec_raw <-
-      try(data_decrypt(decoded_string
-                       , keyAsRaw
-                       , hash(charToRaw("nounce"), size = 24))
-          , silent = TRUE)
+    string_dec_raw <- try(data_decrypt(decoded_string
+                                       , keyAsRaw
+                                       , hash(charToRaw("nounce"), size = 24))
+                          , silent = TRUE)
+
+  } else {
+    string_dec_raw <- try(auth_decrypt(decoded_string
+                                       , key
+                                       , pkey
+                                       , hash(charToRaw("nounce"), size = 24))
+                          , silent = TRUE)
   }
 
   if(is.error(string_dec_raw)){
-    stop("Unable to decrypt. Check whether the input was generated by 'encrypt_file' function. Check whether 'key' and 'method' are correct.")
+    stop("Unable to decrypt. Check whether the input was generated by 'encrypt_file' function.")
   }
 
   # write      ----