+++ R/save.R
@@ -1,25 +1,27 @@
-#' @title Save a R object to a connection
-#'
-#' @aliases save2
+#' @title Save an object to a connection(or a file)
 #' @description \code{save_object} encrypts a R object to raw or text connection
 #'   or a file. \code{retrieve_object} decrypts a raw or a text connection or a
 #'   file (encrypted by \code{save_object}). Note that \code{retrieve_object}
 #'   returns the object.
 #'
 #' @param object A R object to be encrypted
+#' @param key For symmetric encryption, 'pkey' should be NULL (default) and
+#'   'key' can be either a string (Default is 'pass') or a raw object. For
+#'   asymmetric encryption, both 'key' (private key of the encrypter) and 'pkey'
+#'   (public key of the decrypter) should be raw objects.
+#' @param pkey See 'key'
+#' @param ascii TRUE, if the encrypted output is a string(written to the text
+#'   connection). FALSE, if the encrypted output is a raw object(written to the
+#'   raw connection)
 #' @param conn A connection or a file where the encrypted content is written. If
 #'   \code{ascii} is TRUE, an encrypted text is written to the connection. Else,
 #'   when \code{ascii} is FALSE(default), a raw object is written to the
 #'   connection
-#' @param ascii TRUE, if the encrypted output is a string(written to the text
-#'   connection). FALSE, if the encrypted output is a raw object(written to the
-#'   raw connection)
-#' @param key A string without embbeded NULL. Default is 'pass'
-#' @param method Currently, a stub. It should be 'symmetric'(default)
 #'
 #' @return An invisible TRUE
 #'
 #' @examples
+#' # symmetric case:
 #' all(
 #'   save_object(iris, conn = "iris_safer.bin")
 #'   , identical(retrieve_object(conn = "iris_safer.bin"), iris)
@@ -32,13 +34,22 @@
 #'   , unlink("iris_safer_2.txt") == 0
 #' )
 #'
+#' # asymmetric case:
+#' alice <- keypair()
+#' bob   <- keypair()
+#' all(
+#'   save_object(iris, alice$private_key, bob$public_key, conn = "iris_safer.bin")
+#'   , identical(retrieve_object(conn = "iris_safer.bin", bob$private_key, alice$public_key), iris)
+#'   , unlink("iris_safer.bin") == 0
+#' )
+#'
 #' @export
 #'
 save_object     <- function(object
-                            , conn
-                            , ascii  = FALSE
                             , key    = "pass"
-                            , method = "symmetric"){
+                            , pkey   = NULL
+                            , ascii  = FALSE
+                            , conn){
   # assertions                           ----
   assert_that(!missing(object))
   assert_that(is.flag(ascii))
@@ -46,8 +57,14 @@ save_object     <- function(object
   if(is.string(conn)){
     assert_that(!file.exists(conn))
     }
-  assert_that(is.string(key))
-  assert_that(method %in% c("symmetric"))
+  if(is.null(pkey)){
+    method <- "symmetric"
+    assert_that(is.string(key) || is.raw(key))
+  } else {
+    method <- "asymmetric"
+    assert_that(is.raw(key))
+    assert_that(is.raw(pkey))
+  }
 
   # try to see if object is serializable ----
   serializedObject <- try(serialize(object, NULL), silent = TRUE)
@@ -55,17 +72,25 @@ save_object     <- function(object
     stop("Unable to serialize 'object' using 'serialize' function")
   }
 
-  keyAsRaw <- try(hash(charToRaw(key)), silent = TRUE)
-  if(is.error(keyAsRaw)){
-    stop("Unable to convert 'key' into raw. Possibly encountered an embedded NULL.")
+  if(is.string(key)){
+    keyAsRaw <- try(hash(charToRaw(key)), silent = TRUE)
+    if(is.error(keyAsRaw)){
+      stop("Unable to convert 'key' into raw. Possibly encountered an embedded NULL.")
+    }
+  } else {
+    keyAsRaw <- key
   }
 
   # encrypt raw data                     ----
   if(method == "symmetric"){
     object_enc_raw <- data_encrypt(serializedObject
                                    , keyAsRaw
-                                   , hash(charToRaw("nounce"), size = 24)
-    )
+                                   , hash(charToRaw("nounce"), size = 24))
+  } else {
+    object_enc_raw <- auth_encrypt(serializedObject
+                                   , key
+                                   , pkey
+                                   , hash(charToRaw("nounce"), size = 24))
   }
   attributes(object_enc_raw) <- NULL
 
@@ -80,7 +105,7 @@ save_object     <- function(object
              )
       } else {
         on.exit(close(conn), add = TRUE)
-        }
+      }
     }
 
     # try to write text file
@@ -107,26 +132,32 @@ save_object     <- function(object
       stop("Unable to write to the connection or file. Ensure that 'conn' is a open writable raw connection")
     }
   }
-  return(TRUE)
+  return(invisible(TRUE))
 }
 
-#' @title Retrieve a R object from a connection
-#'
-#' @aliases load2
+#' @title Retrieve an object from a connection(or a file)
 #' @description \code{save_object} encrypts a R object to raw or text connection
 #'   or a file. \code{retrieve_object} decrypts a raw or a text connection or a
 #'   file (encrypted by \code{save_object}). Note that \code{retrieve_object}
 #'   returns the object.
 #'
-#' @param conn A connection or a file to be decrypted.
-#' @param ascii TRUE, if conn is a text connection or a ascii file. FALSE, if
-#'   conn is a raw connection or a binary file
-#' @param key A string without embbeded NULL. Default is 'pass'.
-#' @param method Currently, a stub. It should be 'symmetric'(default).
+#' @param conn A connection or a file where the decrypted content is written. If
+#'   \code{ascii} is TRUE, an decrypted text is written to the connection. Else,
+#'   when \code{ascii} is FALSE(default), a raw object is written to the
+#'   connection
+#' @param key For symmetric decryption, 'pkey' should be NULL (default) and
+#'   'key' can be either a string (Default is 'pass') or a raw object. For
+#'   asymmetric decryption, both 'key' (private key of the decrypter) and 'pkey'
+#'   (public key of the encrypter) should be raw objects.
+#' @param pkey See 'key'
+#' @param ascii TRUE, if the encrypted output is a string(written to the text
+#'   connection). FALSE, if the encrypted output is a raw object(written to the
+#'   raw connection)
 #'
 #' @return An invisible TRUE
 #'
 #' @examples
+#' # symmetric case:
 #' all(
 #'   save_object(iris, conn = "iris_safer.bin")
 #'   , identical(retrieve_object(conn = "iris_safer.bin"), iris)
@@ -139,20 +170,35 @@ save_object     <- function(object
 #'   , unlink("iris_safer_2.txt") == 0
 #' )
 #'
+#' # asymmetric case:
+#' alice <- keypair()
+#' bob   <- keypair()
+#' all(
+#'   save_object(iris, alice$private_key, bob$public_key, conn = "iris_safer.bin")
+#'   , identical(retrieve_object(conn = "iris_safer.bin", bob$private_key, alice$public_key), iris)
+#'   , unlink("iris_safer.bin") == 0
+#' )
+#'
 #' @export
 #'
 retrieve_object <- function(conn
-                            , ascii  = FALSE
-                            , key    = "pass"
-                            , method = "symmetric"){
+                            , key   = "pass"
+                            , pkey  = NULL
+                            , ascii = FALSE){
   # assertions                     ----
   if(!is.string(conn)){
     assert_that("connection" %in% class(conn))
   } else {
     assert_that(file.exists(conn))
   }
-  assert_that(is.string(key))
-  assert_that(method %in% c("symmetric"))
+  if(is.null(pkey)){
+    method <- "symmetric"
+    assert_that(is.string(key) || is.raw(key))
+  } else {
+    method <- "asymmetric"
+    assert_that(is.raw(key))
+    assert_that(is.raw(pkey))
+  }
 
   # try to read serialized content ----
   if(ascii){
@@ -174,7 +220,7 @@ retrieve_object <- function(conn
     encryptedObject <- try(base64decode(encryptedText), silent = TRUE)
 
     if(is.error(encryptedObject)){
-      stop("Unable to decode the encrypted text. Ensure that connection or the file was written by 'save_object' or 'save2' function")
+      stop("Unable to decode the encrypted text. Ensure that connection or the file was written by 'save_object' function.")
     }
   } else { # binary case
 
@@ -190,11 +236,20 @@ retrieve_object <- function(conn
     encryptedObject <- try(unserialize(conn), silent = TRUE)
 
     if(is.error(encryptedObject)){
-      stop("Unable to 'unserialize' the connection or file. Ensure that the connection is open raw connection. Ensure that connection or the file was written by 'save_object' or 'save2' function")
+      stop("Unable to 'unserialize' the connection or file. Ensure that the connection is open raw connection. Ensure that connection or the file was written by 'save_object' function.")
     }
   }
 
-  keyAsRaw <- try(hash(charToRaw(key)), silent = TRUE)
+  if(is.string(key)){
+    keyAsRaw <- try(hash(charToRaw(key)), silent = TRUE)
+  } else {
+    keyAsRaw <- key
+  }
+
+  if(is.error(keyAsRaw)){
+    stop("Unable to convert 'key' into raw. Possibly encountered an embedded NULL.")
+  }
+
   if(is.error(keyAsRaw)){
     stop("Unable to convert 'key' into raw. Possibly encountered an embedded NULL.")
   }
@@ -203,13 +258,20 @@ retrieve_object <- function(conn
   if(method == "symmetric"){
     object_dec_raw <-
       try(data_decrypt(encryptedObject
-                       , hash(charToRaw(key))
+                       , keyAsRaw
+                       , hash(charToRaw("nounce"), size = 24))
+          , silent = TRUE)
+  } else {
+    object_dec_raw <-
+      try(auth_decrypt(encryptedObject
+                       , key
+                       , pkey
                        , hash(charToRaw("nounce"), size = 24))
           , silent = TRUE)
   }
 
   if(is.error(object_dec_raw)){
-    stop("Unable to decrypt. Check whether the input was generated by 'save_object' or 'save2' function. Check whether 'key' and 'method' are correct.")
+    stop("Unable to decrypt. Check whether the input was generated by 'save_object' function.")
   }
 
   # return                         ----