+++ NEWS.md
@@ -1,1 +1,4 @@
-- safer first release
+- safer(0.2.0)
+    - added 'asymmetric' encryption option
+
+- safer (0.1.0) first release